requirements.txt
lxml==4.9.2
python_docx==1.1.2
pywin32==305
